Module Mirage_kv

MirageOS signatures for key/value devices

v4.0.1

Mirage_kv

MirageOS key-value stores are nested dictionaries, associating structured keys to either dictionaries or values.

module Key : sig ... end

type key = Key.t

The type for keys.

type error = [

| `Not_found of key	(*
key not found
*)

| `Dictionary_expected of key	(*
key does not refer to a dictionary.
*)

| `Value_expected of key	(*
key does not refer to a value.
*)

]

The type for errors.

val pp_error : error Fmt.t

pp_error is the pretty-printer for errors.

module type RO = sig ... end

type write_error = [

| error

| `No_space	(*
No space left on the device.
*)

| `Too_many_retries of int	(*
batch has been trying to commit n times without success.
*)

]

val pp_write_error : write_error Fmt.t

pp_write_error is the pretty-printer for write errors.

module type RW = sig ... end